Move all tests into test module
