+++ src/lib.rs
@@ -125,7 +125,7 @@ impl Authorizer {
                 }));
             }
 
-            // Verify that the capability doesn't break out of capabilitys:
+            // Verify that the capability doesn't break out of capabilities:
             if !permitted(original) {
                 return Err(e!(InvocationError::NotPermitted));
             }
@@ -637,225 +637,4 @@ fn base32_bytes(s: &str) -> Result<Vec<u8>, DecodeError> {
 }
 
 #[cfg(test)]
-mod tests {
-    use ed25519_dalek::SIGNATURE_LENGTH;
-
-    use super::*;
-
-    #[derive(Debug, Clone, Copy, PartialEq, Eq, Serialize, Deserialize)]
-    pub(crate) enum Rpc {
-        Read,
-        ReadWrite,
-        All,
-    }
-
-    impl Capability for Rpc {
-        fn permits(&self, other: &Self) -> bool {
-            match (self, other) {
-                (Rpc::All, _) => true,
-                (Rpc::ReadWrite, Rpc::Read | Rpc::ReadWrite) => true,
-                (Rpc::ReadWrite, _) => false,
-                (Rpc::Read, Rpc::Read) => true,
-                (Rpc::Read, _) => false,
-            }
-        }
-    }
-
-    pub(crate) fn key(byte: u8) -> SigningKey {
-        SigningKey::from_bytes(&[byte; 32])
-    }
-
-    #[test]
-    fn v2_roundtrip() {
-        let issuer = key(0);
-        let audience = key(1).verifying_key();
-        let delegation: OpaqueDelegation =
-            Delegation::issuing_builder(&issuer, audience, &Rpc::ReadWrite)
-                .sign(Expires::Never)
-                .into();
-
-        let bytes = delegation.encode();
-        assert_eq!(bytes[0], 2);
-        let decoded = OpaqueDelegation::decode(&bytes).unwrap();
-        assert_eq!(decoded, delegation);
-        let decoded = Delegation::<Rpc>::decode(&bytes).unwrap();
-        assert_eq!(decoded.into_opaque(), delegation);
-
-        assert_eq!(delegation.issuer(), &issuer.verifying_key());
-        assert_eq!(delegation.audience(), &audience);
-        assert_eq!(delegation.expires(), &Expires::Never);
-        assert_eq!(delegation.capability_issuer(), &issuer.verifying_key());
-        assert_eq!(delegation.capability(), &[1]);
-
-        let wire = postcard::to_stdvec(&delegation).unwrap();
-        assert!(wire.ends_with(&bytes));
-        let deserialized: OpaqueDelegation = postcard::from_bytes(&wire).unwrap();
-        assert_eq!(deserialized, delegation);
-    }
-
-    #[test]
-    fn v2_decode_rejects_tampering() {
-        let issuer = key(0);
-        let audience = key(1).verifying_key();
-        let delegation: OpaqueDelegation =
-            Delegation::issuing_builder(&issuer, audience, &Rpc::Read)
-                .sign(Expires::Never)
-                .into();
-        let good = delegation.encode();
-
-        let mut forged = good.clone();
-        let n = forged.len();
-        forged[n - SIGNATURE_LENGTH..].fill(0);
-        assert!(OpaqueDelegation::decode(&forged).is_err());
-
-        let mut widened = good.clone();
-        let capability_offset = 1 + 32 + 32 + 1 + 1 + 1;
-        assert_eq!(widened[capability_offset - 1], 1);
-        assert_eq!(widened[capability_offset], 0);
-        widened[capability_offset] = 2;
-        assert!(OpaqueDelegation::decode(&widened).is_err());
-
-        let mut padded = good.clone();
-        let signature_start = padded.len() - SIGNATURE_LENGTH;
-        padded.insert(signature_start, 0);
-        assert!(OpaqueDelegation::decode(&padded).is_err());
-
-        let mut versioned = good.clone();
-        versioned[0] = 3;
-        assert!(OpaqueDelegation::decode(&versioned).is_err());
-
-        assert!(OpaqueDelegation::decode(&[]).is_err());
-    }
-
-    #[test]
-    fn two_link_chain_invocation() {
-        let service = key(0);
-        let alice = key(1);
-        let bob = key(2);
-
-        let root: OpaqueDelegation =
-            Delegation::issuing_builder(&service, alice.verifying_key(), &Rpc::All)
-                .sign(Expires::At(4_102_444_800))
-                .into();
-
-        let link: OpaqueDelegation = Delegation::delegating_builder(
-            &alice,
-            bob.verifying_key(),
-            service.verifying_key(),
-            &Rpc::Read,
-        )
-        .sign(Expires::Never)
-        .into();
-
-        let authorizer = Authorizer::new(service.verifying_key());
-        let chain = [&root, &link];
-        let now = SystemTime::UNIX_EPOCH + Duration::from_secs(1_700_000_000);
-
-        authorizer
-            .check_opaque_invocation_from_at(now, bob.verifying_key(), Rpc::Read, &chain)
-            .unwrap();
-        assert!(authorizer
-            .check_opaque_invocation_from_at(now, bob.verifying_key(), Rpc::ReadWrite, &chain)
-            .is_err());
-        assert!(authorizer
-            .check_opaque_invocation_from_at(now, key(3).verifying_key(), Rpc::Read, &chain)
-            .is_err());
-        let late = SystemTime::UNIX_EPOCH + Duration::from_secs(4_102_444_801);
-        assert!(authorizer
-            .check_opaque_invocation_from_at(late, bob.verifying_key(), Rpc::Read, &chain)
-            .is_err());
-    }
-
-    #[test]
-    fn chain_must_start_at_the_authorizer() {
-        let service = key(0);
-        let alice = key(1);
-        let bob = key(2);
-
-        let alice_grant: OpaqueDelegation =
-            Delegation::issuing_builder(&alice, bob.verifying_key(), &Rpc::All)
-                .sign(Expires::Never)
-                .into();
-        let authorizer = Authorizer::new(service.verifying_key());
-        assert!(authorizer
-            .check_opaque_invocation_from(bob.verifying_key(), Rpc::Read, &[&alice_grant])
-            .is_err());
-    }
-
-    #[test]
-    fn subject_must_be_the_authorizer() {
-        let service = key(0);
-        let other = key(1);
-        let alice = key(2);
-
-        let delegation: OpaqueDelegation = Delegation::delegating_builder(
-            &service,
-            alice.verifying_key(),
-            other.verifying_key(),
-            &Rpc::All,
-        )
-        .sign(Expires::Never)
-        .into();
-        let authorizer = Authorizer::new(service.verifying_key());
-        assert!(authorizer
-            .check_opaque_invocation_from(alice.verifying_key(), Rpc::Read, &[&delegation])
-            .is_err());
-    }
-
-    #[test]
-    fn typed_delegation() {
-        let issuer = key(0);
-        let audience = key(1).verifying_key();
-
-        let typed =
-            Delegation::issuing_builder(&issuer, audience, &Rpc::ReadWrite).sign(Expires::Never);
-        assert_eq!(typed.capability(), Rpc::ReadWrite);
-
-        let untyped: OpaqueDelegation = typed.clone().into();
-        let again = Delegation::<Rpc>::try_from(untyped.clone()).unwrap();
-        assert_eq!(again, typed);
-
-        #[derive(Debug, Serialize, Deserialize)]
-        struct OtherVocabulary {
-            topic: String,
-            write: bool,
-        }
-        assert!(Delegation::<OtherVocabulary>::try_from(untyped.clone()).is_err());
-
-        let wire = postcard::to_stdvec(&typed).unwrap();
-        assert_eq!(wire, postcard::to_stdvec(&untyped).unwrap());
-
-        let ok: Delegation<Rpc> = postcard::from_bytes(&wire).unwrap();
-        assert_eq!(ok, typed);
-        assert!(postcard::from_bytes::<Delegation<OtherVocabulary>>(&wire).is_err());
-
-        let bob = key(2);
-        let typed_link = Delegation::delegating_builder(
-            &key(1),
-            bob.verifying_key(),
-            issuer.verifying_key(),
-            &Rpc::Read,
-        )
-        .sign(Expires::Never);
-        let root = Delegation::issuing_builder(&issuer, audience, &Rpc::All).sign(Expires::Never);
-        let authorizer = Authorizer::new(issuer.verifying_key());
-        authorizer
-            .check_invocation_from(bob.verifying_key(), Rpc::Read, &[&root, &typed_link])
-            .unwrap();
-        assert!(authorizer
-            .check_invocation_from(bob.verifying_key(), Rpc::ReadWrite, &[&root, &typed_link])
-            .is_err());
-    }
-
-    #[test]
-    fn owner_needs_no_chain() {
-        let service = key(0);
-        let authorizer = Authorizer::new(service.verifying_key());
-        authorizer
-            .check_opaque_invocation_from(service.verifying_key(), Rpc::All, &[])
-            .unwrap();
-        assert!(authorizer
-            .check_opaque_invocation_from(key(1).verifying_key(), Rpc::Read, &[])
-            .is_err());
-    }
-}
+mod tests;
